Add an image to header/ Change content and config
Created a header on layout.liquid file with title and html command to
insert image. I configured this command (liquid = `{{ image}}`)
to  chnage image depending on the page (tags). The image file is
set in individual page markdown file.
However, the image wasn't loading because we needed to copy the assets
folder (where the image is) to the output folder. We did this on
`config.js`.
Changed some content in the index, about and first post.