Change template language and add a new template
Changed the template language from Liquid to Nunjucks because of
the lack of the Template Inheritance. This included changes in the
layout file itself (code), name of the file (`layout.njk)`, and name in
the Markdown headers (all the .md files).
Added a new template for the posts-index (`layout-posts.njk`). It uses
the parent template (`layout.njk`). Modifications were made based on
the "blocks" definition
(https://mozilla.github.io/nunjucks/templating.html#template-inheritance).